Cambios en Contratos y Cotizaciones

+++ ERPMVC/Models/Clientes/CustomerContractLine.cs
@@ -40,7 +40,7 @@ namespace ERPMVC.Models
 
         public string TipoCobroName { get; set; }
 
-        public int PeriodoCobro { get; set; }
+        public decimal PeriodoCobro { get; set; }
 
     }
 }
